Split SN-client objects file
